Move master file watcher to make it available to other modules
This commit moves the new master node file watcher to the
o.e.common.file package so that it can be accessible to existing uses of
AbstractFileWatchingService.
